+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.2.18+1] - October 28th, 2020
 
-* Added `appIdentifier` to `ExternalTestDriver`.
+* Added `appIdentifier` to `ExternalTestDriver` and `GoldenTestImages` id.
+* Added `TestAppSettings`.
 
 
 ## [1.2.18] - October 28th, 2020